Remove apply(); rename set/unset to attach/detach
apply() is confusing and interferes with the viewport in unexpected ways. It's not clear what the viewport coordinates should be, if the coordinate system is moved as apply() expects.

The user can still use love.graphics.applyTransform(cam:getTransform()) and handle the viewport in the preferred way.

As for the renaming, both hump.camera and STALKER-X use these names.